requirements.txt
pypdf
bcrypt
langchain
Flask
Flask-Cors
together
pydantic==1.10.9
langchain
openai
faiss-cpu
python-dotenv
pinecone-client
psycopg2-binary
PyJWT
requests
SQLAlchemy-Utils
flask_jwt_extended
PyPDF2
PyMuPDF
Pillow
tiktoken
python-docx
